[ebpf/manager] Resolve ELF symbol from memory (/proc/pid/maps) if the library file is not accessible/available
FindSymbolOffsets() will try to extract symbols from the library ELF file.
If it can't fallback to parse the ELF available from memory (proc/pid/maps)
But the original ELF sections could not be avaiable (not mapped in memory)
so the last chance is to get the table file offset (DYNTAB, SYMTAB)
from the prog ELF section DT_DYNAMIC and reconstruct ELF section
to extract dynamic symbols.

OpenAndListSymbols(path) : read the ELF thanks to debug/elf

OpenAndListSymbolsFromPID(pid, path) : read the ELF thanks to elfmap

If the .so is not accessible, elfmap would ptrace() PEEK_DATA and copy the mapped lib.
If the caller lookup to himself, so the same memspace, we just memcpy the sections.